+++ backend/users/src/models/mod.rs
@@ -41,12 +41,12 @@ impl FromSql<Text, Pg> for Role {
     }
 }
 
-#[derive(Queryable, Selectable)]
+#[derive(Queryable, Selectable, AsChangeset, Identifiable)]
 #[diesel(table_name = utils::db::schema::users)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 #[derive(Debug, Serialize, Deserialize)]
 pub struct User {
-    pub id: i32,
+    pub id: i64,
     pub name: Option<String>,
     pub give_name: Option<String>,
     pub email: Option<String>,
